Remove unused library.
I was experimenting with web components but it didn't pan out. Something
about it not being able to hotload on code change made me not want to
use it for now. Maybe it's the library and maybe it's the implementation
of web components in Chrome but I'd rather not have some distraction for
building this app for now.